Move the validator code in approllout to own file
This follows the example from the chart assignment controller.
